backend/requirements-py36.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
gunicorn==20.1.0
sqlalchemy==1.4.23
asyncpg==0.24.0
alembic==1.7.3
psycopg2-binary==2.9.1
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
redis==3.5.3
python-dotenv==0.19.0
python-multipart==0.0.5
